chore(typescript): config app to use typescript
expo expects the entry point to be App.js, thus it imports the
App from the src directory then exports it as it's own module.
The rest of the app is then written in typescript files using the
tsx extension.
expo also needed to be configured with packagerOpts in app.json